refactor(core): move setObservableConfig to separate file
