+++ docs/guides/interop.md
@@ -9,13 +9,52 @@
 
 <!-- END doctoc generated TOC please keep comment here to allow auto update -->
 
-Stapp uses a reactive approach to business logic handling. Observables, streams, and epics are the core parts of Stapp.
+Stapp uses a reactive approach to business logic handling. Observables, streams, and epics are the core parts of Stapp. 
 
-Since version 2.0.0 we decided to move from RxJS to our own incredibly light observable implementation. We believe that using RxJS imposes too many restrictions on Stapp end-users. RxJS is standard de-facto but is far not the only reactive library out the wild.
-
-Since version 2.0.0 epics receive a stream of events and a stream of state changes based on `light-observable`, the standard implementation of the [Observable proposal](https://github.com/tc39/proposal-observable).
+Observables in Stapp conform to the [Observable proposal](https://github.com/tc39/proposal-observable). See [`light-observable`](http://light-observable.js.org/) for more details.
 
 This means that those streams are fully compatible with any ECMAScript compatible reactive library, including **RxJS**, **most**, **kefir**, **bacon** and so on.
+
+There are several ways to use these streams with preferable reactive library.
+
+## Global config: `setObservableConfig()`
+```typescript
+type setObservableConfig<Stream> = ({
+  fromESObservable: (observable: Observable<any>) => Stream,
+  toESObservable: (stream: Stream) => Observable<any>
+})
+```
+**Note: setObservableConfig() uses global state, and could break apps if used inside a package intended to be shared.**
+
+### Example
+```js
+import { from } from 'rxjs'
+import { setObservableConfig } from 'stapp'
+
+setObservableConfig({
+  // Converts a plain ES observable to an RxJS 5 observable
+  fromESObservable: Rx.Observable.from
+})
+```
+
+```javascript
+import { from } from 'rxjs'
+import { switchMap, filter } from 'rxjs/operators'
+
+const module = {
+  name: 'my module',
+  epic: (event$) => event$.pipe( // `pipe` is the only non-standard method in `light-observable`
+    from, // here we transform es observable to Rx observable
+    filter(({ type }) => type === 'some event'),
+    switchMap(({ payload }) => myEffect(payload))
+  )
+}
+```
+This is fine, but doing this many times may be too verbose, especially when there are many epics in a module.
+
+
+## `setObservableConfig`
+**Note: setObservableConfig() uses global state, and could break apps if used inside a package intended to be shared.**
 
 ## RxJS
 `light-observable` stream can be used with RxJS easily. Just wrap it with `from`:
