+++ packages/stapp/src/stapp.ts
@@ -1,10 +1,11 @@
 // Core utils
-export { createApp, setObservableConfig } from './core/createApp/createApp'
+export { createApp } from './core/createApp/createApp'
 export { createEvent } from './core/createEvent/createEvent'
 export { createEffect } from './core/createEffect/createEffect'
 export { createReducer } from './core/createReducer/createReducer'
 
 // Epic utils
+export { setObservableConfig } from './core/createApp/setObservableConfig'
 export { select, selectArray } from './helpers/select/select'
 export { getEventType } from './helpers/getEventType/getEventType'
 export { combineEpics } from './helpers/combineEpics/combineEpics'
